Refactor JobSkills
